requirements.txt
python-dotenv==1.0.0
requests==2.31.0
langchain-core>=0.1.8,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain>=0.1.0,<0.2.0
